requirements.txt
Flask==0.10.1
Jinja2==2.7.3
MarkupSafe==0.23
PyYAML==3.11
Werkzeug==0.10.1
facebook-sdk==0.4.0
ipython==2.4.0
itsdangerous==0.24
pymongo==2.8
wsgiref==0.1.2
